tree: cast printf args to (unsigned int) on uint32_t/data_t/offs_t
On toolchains where uint32_t = unsigned long int (3DS CTR, some newlib
arm targets), %x / %X / %d / %u format specifiers warn or error
out because they expect "unsigned int", not "unsigned long".  On
x86-64 LP64 Linux the conflict is latent (uint32_t = unsigned int).

These are all in non-hot debug / disassembly / log paths where the
ABI cost of the explicit cast is zero (the original sites passed
unsigned int -size values via implicit promotion, and the cast just
makes that explicit to the type-checker).

Files touched:
- src/common.c              ROM CRC dump
- src/unzip.c               unzip log (CRC of entries)
- src/sndhrdw/exidy440.c    sound chip volume log
- src/drivers/88games.c     debug printout
- src/drivers/tnzs.c        voice debug printout
- src/vidhrdw/{gsword,kyugo,mitchell}.c  videoctrl debug
- src/cpu/m6800/m6800.c     m6800 disassembler (6 sites)
- src/cpu/arm/arm.c         arm disassembler
- src/cpu/adsp2100/adsp2100.c  adsp2100 disassembler (2 sites)
- src/cpu/t11/t11.c         t11 disassembler
- src/cpu/tms9900/99xxcore.h  tms9900 disassembler (in shared header
                              included by both tms9980a.c)

No behavior change on any host; clears all "-Wformat=" warnings under
the 3DS / newlib int32_t = long simulation.

+++ src/common.c
@@ -152,7 +152,7 @@ int readroms(void)
 					/* NS981003: support for "load by CRC" */
 					char crc[9];
 
-					sprintf(crc,"%08x",romp->crc);
+					sprintf(crc,"%08x",(unsigned int)romp->crc);
 					drv = Machine->gamedrv;
 					do
 					{
