+++ src/vidhrdw/mitchell.c
@@ -177,7 +177,7 @@ WRITE_HANDLER( pang_gfxctrl_w )
 //logerror("PC %04x: pang_gfxctrl_w %02x\n",cpu_get_pc(),data);
 {
 	char baf[40];
-	sprintf(baf,"%02x",data);
+	sprintf(baf,"%02x",(unsigned int)data);
 //	usrintf_showmessage(baf);
 }
